+++ packages/polaris-viz/src/components/Annotations/components/AnnotationLabel/AnnotationLabel.tsx
@@ -3,7 +3,7 @@ import {LINE_HEIGHT, useTheme} from '@shopify/polaris-viz-core';
 
 import {useBrowserCheck} from '../../../../hooks/useBrowserCheck';
 import {SingleTextLine} from '../../../Labels';
-import {PILL_HEIGHT, PILL_OPACITY, PILL_PADDING} from '../../constants';
+import {PILL_HEIGHT, PILL_PADDING} from '../../constants';
 import type {AnnotationPosition} from '../../types';
 
 import styles from './AnnotationLabel.scss';
@@ -58,7 +58,7 @@ export function AnnotationLabel({
         width={width}
         fill={selectedTheme.annotations.backgroundColor}
         ry={PILL_HEIGHT / 2}
-        opacity={isFirefox ? 1 : PILL_OPACITY}
+        opacity={isFirefox ? 1 : selectedTheme.annotations.pillOpacity}
       />
 
       <SingleTextLine